test(zoo): combine tests for `.core.zoo` into one file
